move getHouse houseActions instead of necessityActions
